자격증 이력 코드 리팩토링
