+++ src/main/java/com/nexus/sion/feature/member/command/application/service/DeveloperCertificateHistoryService.java
@@ -1,0 +1,7 @@
+package com.nexus.sion.feature.member.command.application.service;
+
+import com.nexus.sion.feature.member.command.application.dto.request.UserCertificateHistoryRequest;
+
+public interface DeveloperCertificateHistoryService {
+    void registerUserCertificate(String employeeId, UserCertificateHistoryRequest request);
+}
